@@ -13,24 +13,26 @@
         }
     },
     "require": {
-        "php": ">=5.3.23",
+        "php": ">=5.5",
         "zendframework/zend-stdlib": "~2.5"
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5-dev",
-            "dev-develop": "2.6-dev"
+            "dev-master": "2.6-dev",
+            "dev-develop": "3.0-dev"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "ZendTest\\EventManager\\": "test/"
+            "ZendTest\\EventManager\\": "test/",
+            "ZendBench\\EventManager\\": "benchmarks/"
         }
     },
     "require-dev": {
         "fabpot/php-cs-fixer": "1.7.*",
-        "phpunit/PHPUnit": "~4.0"
+        "phpunit/PHPUnit": "~4.0",
+        "athletic/athletic": "dev-master"
     }
 }
