@@ -13,13 +13,15 @@
         }
     },
     "require": {
-        "php": ">=5.3.23"
+        "php": "^5.5 || ^7.0",
+        "container-interop/container-interop": "~1.0"
     },
     "require-dev": {
         "zendframework/zend-di": "~2.5",
         "zendframework/zend-mvc": "~2.5",
         "fabpot/php-cs-fixer": "1.7.*",
-        "phpunit/PHPUnit": "~4.0"
+        "phpunit/PHPUnit": "~4.0",
+        "athletic/athletic": "dev-master"
     },
     "suggest": {
         "ocramius/proxy-manager": "ProxyManager 0.5.* to handle lazy initialization of services",
@@ -29,13 +31,14 @@
     "prefer-stable": true,
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5-dev",
-            "dev-develop": "2.6-dev"
+            "dev-master": "2.7-dev",
+            "dev-develop": "3.0-dev"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "ZendTest\\ServiceManager\\": "test/"
+            "ZendTest\\ServiceManager\\": "test/",
+            "ZendBench\\ServiceManager\\": "benchmarks/"
         }
     }
 }
