@@ -13,7 +13,8 @@
         }
     },
     "require": {
-        "php": ">=5.3.23"
+        "php": ">=5.5",
+        "zendframework/zend-hydrator": "~1.0"
     },
     "require-dev": {
         "zendframework/zend-config": "~2.5",
@@ -23,7 +24,8 @@
         "zendframework/zend-servicemanager": "~2.5",
         "zendframework/zend-filter": "~2.5",
         "fabpot/php-cs-fixer": "1.7.*",
-        "phpunit/PHPUnit": "~4.0"
+        "phpunit/PHPUnit": "~4.0",
+        "athletic/athletic": "~0.1"
     },
     "suggest": {
         "zendframework/zend-eventmanager": "To support aggregate hydrator usage",
@@ -35,13 +37,14 @@
     "prefer-stable": true,
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5-dev",
-            "dev-develop": "2.6-dev"
+            "dev-master": "2.7-dev",
+            "dev-develop": "2.8-dev"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "ZendTest\\Stdlib\\": "test/"
+            "ZendTest\\Stdlib\\": "test/",
+            "ZendBench\\Stdlib\\": "benchmark/"
         }
     }
 }
