@@ -2,14 +2,15 @@
 
 namespace Heise\Shariff\Backend;
 
-use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
+use GuzzleHttp\Exception\TransferException;
 use GuzzleHttp\Pool;
 use Heise\Shariff\CacheInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Log\LoggerInterface;
 
 /**
- * Class BackendManager
- *
- * @package Heise\Shariff\Backend
+ * Class BackendManager.
  */
 class BackendManager
 {
@@ -19,48 +20,74 @@ class BackendManager
     /** @var CacheInterface */
     protected $cache;
 
-    /** @var Client */
+    /** @var ClientInterface */
     protected $client;
 
-    /** @var string */
-    protected $domain;
+    /** @var array */
+    protected $domains = [];
 
     /** @var ServiceInterface[] */
     protected $services;
 
+    /** @var LoggerInterface */
+    protected $logger;
+
     /**
-     * @param string $baseCacheKey
-     * @param CacheInterface $cache
-     * @param Client $client
-     * @param string $domain
+     * @param string             $baseCacheKey
+     * @param CacheInterface     $cache
+     * @param ClientInterface    $client
+     * @param array|string       $domains
      * @param ServiceInterface[] $services
      */
-    public function __construct($baseCacheKey, CacheInterface $cache, Client $client, $domain, array $services)
-    {
+    public function __construct(
+        $baseCacheKey,
+        CacheInterface $cache,
+        ClientInterface $client,
+        $domains,
+        array $services
+    ) {
         $this->baseCacheKey = $baseCacheKey;
         $this->cache = $cache;
         $this->client = $client;
-        $this->domain = $domain;
+        if (is_array($domains)) {
+            $this->domains = $domains;
+        } elseif (is_string($domains)) {
+            trigger_error(
+                'Passing a domain string is deprecated since 5.1, please use an array instead.',
+                E_USER_DEPRECATED
+            );
+            $this->domains = [$domains];
+        }
         $this->services = $services;
     }
 
     /**
      * @param string $url
+     *
      * @return bool
      */
     private function isValidDomain($url)
     {
-        if ($this->domain) {
+        if (!empty($this->domains)) {
             $parsed = parse_url($url);
-            if ($parsed["host"] != $this->domain) {
-                return false;
-            }
+
+            return in_array($parsed['host'], $this->domains, true);
         }
+
         return true;
     }
 
     /**
+     * @param LoggerInterface $logger
+     */
+    public function setLogger(LoggerInterface $logger = null)
+    {
+        $this->logger = $logger;
+    }
+
+    /**
      * @param string $url
+     *
      * @return array|mixed|null
      */
     public function get($url)
@@ -70,7 +97,7 @@ class BackendManager
         $cache_key = md5($url.$this->baseCacheKey);
 
         if (!filter_var($url, FILTER_VALIDATE_URL)) {
-            return null;
+            return;
         }
 
         if ($this->cache->hasItem($cache_key)) {
@@ -78,30 +105,38 @@ class BackendManager
         }
 
         if (!$this->isValidDomain($url)) {
-            return null;
+            return;
         }
 
         $requests = array_map(
             function ($service) use ($url) {
-                /** @var ServiceInterface $service */
+                /* @var ServiceInterface $service */
                 return $service->getRequest($url);
             },
             $this->services
         );
 
+        /** @var ResponseInterface[] $results */
         $results = Pool::batch($this->client, $requests);
 
-        $counts = array();
+        $counts = [];
         $i = 0;
         foreach ($this->services as $service) {
-            if (method_exists($results[$i], "json")) {
-                try {
-                    $counts[ $service->getName() ] = (int)$service->extractCount($results[$i]->json());
-                } catch (\Exception $e) {
-                    // Skip service if broken
+            if ($results[$i] instanceof TransferException) {
+                if ($this->logger !== null) {
+                    $this->logger->warning($results[$i]->getMessage(), ['exception' => $results[$i]]);
                 }
+            } else {
+                try {
+                    $content = $service->filterResponse($results[$i]->getBody()->getContents());
+                    $counts[$service->getName()] = (int) $service->extractCount(json_decode($content, true));
+                } catch (\Exception $e) {
+                    if ($this->logger !== null) {
+                        $this->logger->warning($e->getMessage(), ['exception' => $e]);
+                    }
+                }
+                ++$i;
             }
-            $i++;
         }
 
         $this->cache->setItem($cache_key, json_encode($counts));
