@@ -2,16 +2,15 @@
 
 namespace Heise\Shariff\Backend;
 
+use GuzzleHttp\Psr7;
+
 /**
- * Class GooglePlus
- *
- * @package Heise\Shariff\Backend
+ * Class GooglePlus.
  */
 class GooglePlus extends Request implements ServiceInterface
 {
-
     /**
-     * @return string
+     * {@inheritdoc}
      */
     public function getName()
     {
@@ -19,35 +18,40 @@ class GooglePlus extends Request implements ServiceInterface
     }
 
     /**
-     * @param string $url
-     * @return \GuzzleHttp\Message\Request|\GuzzleHttp\Message\RequestInterface
+     * {@inheritdoc}
      */
     public function getRequest($url)
     {
         $gPlusUrl = 'https://clients6.google.com/rpc?key=AIzaSyCKSbrvQasunBoV16zDH9R33D88CeLr9gQ';
-        $json = array(
+        $json = [
             'method' => 'pos.plusones.get',
             'id'     => 'p',
-            'params' => array(
+            'params' => [
                 'nolog'   => 'true',
                 'id'      => $url,
                 'source'  => 'widget',
                 'userId'  => '@viewer',
-                'groupId' => '@self'
-            ),
+                'groupId' => '@self',
+            ],
             'jsonrpc'    => '2.0',
             'key'        => 'p',
-            'apiVersion' => 'v1'
-        );
-        return $this->createRequest($gPlusUrl, 'POST', array('json' => $json));
+            'apiVersion' => 'v1',
+        ];
+
+        $body = Psr7\stream_for(json_encode($json));
+
+        return new \GuzzleHttp\Psr7\Request('POST', $gPlusUrl, [], $body);
     }
 
     /**
-     * @param array $data
-     * @return int
+     * {@inheritdoc}
      */
     public function extractCount(array $data)
     {
-        return $data['result']['metadata']['globalCounts']['count'];
+        if (!empty($data['result']['metadata']['globalCounts']['count'])) {
+            return $data['result']['metadata']['globalCounts']['count'];
+        }
+
+        return 0;
     }
 }
