@@ -2,25 +2,24 @@
 
 namespace Heise\Shariff\Backend;
 
-use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
+use Psr\Http\Message\RequestInterface;
 
 /**
- * Class Request
- *
- * @package Heise\Shariff\Backend
+ * Class Request.
  */
 abstract class Request
 {
-    /** @var Client */
+    /** @var ClientInterface */
     protected $client;
 
     /** @var array */
     protected $config;
 
     /**
-     * @param Client $client
+     * @param ClientInterface $client
      */
-    public function __construct(Client $client)
+    public function __construct(ClientInterface $client)
     {
         $this->client = $client;
     }
@@ -28,21 +27,27 @@ abstract class Request
     /**
      * @param string $url
      * @param string $method
-     * @param array $options
-     * @return \GuzzleHttp\Message\Request
+     * @param array  $options This parameter is ignored and is only present for backwards compatibility reasons.
+     *
+     * @return RequestInterface
+     *
+     * @deprecated This method is not used anymore and will be removed with version 6.
+     *             Use \GuzzleHttp\Psr7\Request directly instead.
      */
-    protected function createRequest($url, $method = 'GET', $options = array())
+    protected function createRequest($url, $method = 'GET', $options = [])
     {
-        // $defaults = array('future' => true, 'debug' => true);
-        $defaults = array('future' => true, 'timeout' => 5.0);
+        trigger_error('This method is not used anymore and will be removed with version 6.'
+                    .' Use \GuzzleHttp\Psr7\Request directly instead.', E_USER_DEPRECATED);
 
-        $req = $this->client->createRequest(
-            $method,
-            $url,
-            array_merge($defaults, $options)
-        );
+        return new \GuzzleHttp\Psr7\Request($method, $url);
+    }
 
-        return $req;
+    /**
+     * {@inheritdoc}
+     */
+    public function filterResponse($content)
+    {
+        return $content;
     }
 
     /**
