@@ -2,21 +2,23 @@
 
 namespace Heise\Shariff\Backend;
 
+use Psr\Http\Message\RequestInterface;
+
 /**
- * Interface ServiceInterface
- *
- * @package Heise\Shariff\Backend
+ * Interface ServiceInterface.
  */
 interface ServiceInterface
 {
     /**
      * @param string $url
-     * @return \GuzzleHttp\Message\Request
+     *
+     * @return RequestInterface
      */
     public function getRequest($url);
 
     /**
      * @param array $data
+     *
      * @return int
      */
     public function extractCount(array $data);
@@ -27,8 +29,14 @@ interface ServiceInterface
     public function getName();
 
     /**
+     * @param string $content
+     *
+     * @return string
+     */
+    public function filterResponse($content);
+
+    /**
      * @param array $config
-     * @return void
      */
     public function setConfig(array $config);
 }
