@@ -3,14 +3,12 @@
 namespace Heise\Shariff\Backend;
 
 /**
- * Class StumbleUpon
- *
- * @package Heise\Shariff\Backend
+ * Class StumbleUpon.
  */
 class StumbleUpon extends Request implements ServiceInterface
 {
     /**
-     * @return string
+     * {@inheritdoc}
      */
     public function getName()
     {
@@ -18,20 +16,21 @@ class StumbleUpon extends Request implements ServiceInterface
     }
 
     /**
-     * @param string $url
-     * @return \GuzzleHttp\Message\Request|\GuzzleHttp\Message\RequestInterface
+     * {@inheritdoc}
      */
     public function getRequest($url)
     {
-        return $this->createRequest('https://www.stumbleupon.com/services/1.01/badge.getinfo?url='.urlencode($url));
+        return new \GuzzleHttp\Psr7\Request(
+            'GET',
+            'https://www.stumbleupon.com/services/1.01/badge.getinfo?url='.urlencode($url)
+        );
     }
 
     /**
-     * @param array $data
-     * @return int
+     * {@inheritdoc}
      */
     public function extractCount(array $data)
     {
-        return (isset($data['result']['views'])) ? $data['result']['views']+0 : 0;
+        return (isset($data['result']['views'])) ? $data['result']['views'] + 0 : 0;
     }
 }
